+++ 1/spectrum2.py
@@ -26,7 +26,7 @@ x = np.zeros(N)
 
 xFFT1 = fft(x, N)
 
-xFFT2 = xFFT1[0:N/2]
+xFFT2 = xFFT1[0:int(N/2)]
 
 spectrum = np.abs(xFFT2)
