+++ 1/spectrum3.py
@@ -17,7 +17,7 @@ for i in range(1, 10000, 2):
 
 xFFT1 = fft(x, N)
 
-xFFT2 = xFFT1[0:N/2]
+xFFT2 = xFFT1[0:int(N/2)]
 
 spectrum = np.abs(xFFT2)
